Add coverage to basicAuth/Rbac scenario
