--- integration-tests/hivemq-client-smallrye/src/test/java/io/quarkiverse/hivemqclient/test/smallrye/PriceResourceIT.java
@@ -1,7 +1,0 @@
-package io.quarkiverse.hivemqclient.test.smallrye;
-
-import io.quarkus.test.junit.QuarkusIntegrationTest;
-
-@QuarkusIntegrationTest
-public class PriceResourceIT extends PriceResourceTest {
-}
