+++ integration-tests/hivemq-client-smallrye/src/test/java/io/quarkiverse/hivemqclient/test/smallrye/RbacAuthPriceIT.java
@@ -1,0 +1,7 @@
+package io.quarkiverse.hivemqclient.test.smallrye;
+
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+
+@QuarkusIntegrationTest
+public class RbacAuthPriceIT extends RbacAuthPriceTest {
+}
