+++ integration-tests/hivemq-client-smallrye/src/test/java/io/quarkiverse/hivemqclient/test/smallrye/RbacAuthPriceTest.java
@@ -1,0 +1,11 @@
+package io.quarkiverse.hivemqclient.test.smallrye;
+
+import io.quarkiverse.hivemqclient.test.smallrye.resources.HivemqProfiles;
+import io.quarkus.test.junit.QuarkusTest;
+import io.quarkus.test.junit.TestProfile;
+
+@TestProfile(HivemqProfiles.BasicAuth.class)
+@QuarkusTest
+public class RbacAuthPriceTest extends CommonScenarios {
+
+}
