+++ integration-tests/hivemq-client-smallrye/src/test/java/io/quarkiverse/hivemqclient/test/smallrye/resources/NoAuthResources.java
@@ -1,0 +1,4 @@
+package io.quarkiverse.hivemqclient.test.smallrye.resources;
+
+public class NoAuthResources extends CommonResources {
+}
