+++ integration-tests/hivemq-client-smallrye/src/test/resources/hivemq-file-rbac-extension/credentials.xml
@@ -1,0 +1,23 @@
+<?xml version="1.0" encoding="UTF-8" standalone="yes"?>
+<file-rbac>
+    <users>
+        <user>
+            <name>test</name>
+            <password>test</password>
+            <roles>
+                <id>superuser</id>
+            </roles>
+        </user>
+    </users>
+    <roles>
+        <role>
+            <id>superuser</id>
+            <permissions>
+                <permission>
+                    <!-- Allow everything -->
+                    <topic>#</topic>
+                </permission>
+            </permissions>
+        </role>
+    </roles>
+</file-rbac>
